+++ src/my_first_adventure_game/game/levels/__init__.py
@@ -1,4 +1,11 @@
+from my_first_adventure_game.game.levels.clearing_map import create_clearing_map
 from my_first_adventure_game.game.levels.demo_map import create_demo_map
 from my_first_adventure_game.game.levels.game_map import GameMap
+from my_first_adventure_game.game.levels.map_exit import MapExit
 
-__all__ = ["GameMap", "create_demo_map"]
+__all__ = [
+    "GameMap",
+    "MapExit",
+    "create_clearing_map",
+    "create_demo_map",
+]